+++ src/extension.ts
@@ -10,11 +10,11 @@ import ChildProcess = require('child_process');
 import path = require('path');
 
 import { getSearchPathOptions, getConfig, getOverrideOrDefaultConfig, SearchTextHolderReplaceRegex } from './dynamicConfig';
-import { outputError, outputWarn, outputInfo, clearOutputChannel, runCommandInTerminal, outputDebug } from './outputUtils';
+import { outputError, outputWarn, outputInfo, clearOutputChannel, runCommandInTerminal, outputDebug, RunCmdTerminalName, disposeTerminal } from './outputUtils';
 import { FindType, SearchProperty, FileExtensionToConfigExtMap } from './ranker';
 import { checkSearchToolExists, IsWindows } from './checkTool';
 import { getCurrentWordAndText } from './utils';
-import { MyCommandType, runFindingCommand, runFindingCommandByCurrentWord } from './commands';
+import { FindCommandType, runFindingCommand, runFindingCommandByCurrentWord } from './commands';
 
 const GetFileLineTextRegex = new RegExp('(.+?):(\\d+):(.*)');
 
@@ -24,7 +24,7 @@ const NotIgnoreErrorRegex = /^(Matched|Replaced) \d+ .*?(Error|error)/;
 const CheckMaxSearchDepthRegex = /\s+(-k\s*\d+|--max-depth\s+\d+)/;
 
 // Use bytes/second should be more precise.
-const ExpectedMinLinesPerSecond = 30 * 10000;
+const ExpectedMinLinesPerSecond = 20 * 10000;
 const ExpectedMaxTimeCostSecond = 3.0;
 let SearchToCostSumMap = new Map<FindType, Number>();
 let SearchTimesMap = new Map<FindType, Number>();
@@ -32,8 +32,9 @@ let SearchTimesMap = new Map<FindType, Number>();
 // const MsrExePaths = (process.env.Path || '').split(/\s*;\s*/).map(a => path.join(a, 'msr.exe')).filter(a => fs.existsSync(a));
 // const MsrExePath = MsrExePaths.length > 0 ? MsrExePaths[0] : 'msr.exe';
 
+let lastSearchTime = process.hrtime();
 let MyConfig = getConfig();
-let RootConfig = getConfig().RootConfig || vscode.workspace.getConfiguration('msr');
+let RootConfig = MyConfig.RootConfig || vscode.workspace.getConfiguration('msr');
 checkSearchToolExists();
 // vscode.languages.getLanguages().then((languages: string[]) => { console.log("Known languages: " + languages); });
 
@@ -51,73 +52,105 @@ export function activate(context: vscode.ExtensionContext) {
 	// Listening to configuration changes
 	context.subscriptions.push(vscode.workspace.onDidChangeConfiguration(e => {
 		if (e.affectsConfiguration('msr')) {
-			getConfig(true);
+			MyConfig = getConfig(true);
 			RootConfig = vscode.workspace.getConfiguration('msr');
 			outputDebug('msr.enable.definition = ' + RootConfig.get('enable.definition'));
 			outputDebug('msr.enable.reference = ' + RootConfig.get('enable.reference'));
 			outputDebug('msr.enable.findingCommands = ' + RootConfig.get('enable.findingCommands'));
 			outputDebug('msr.quiet = ' + RootConfig.get('quiet'));
 			outputDebug('msr.debug = ' + RootConfig.get('debug'));
+			outputDebug('msr.disable.extensionPatterns = ' + RootConfig.get('disable.extensionPatterns'));
 		}
 	}));
 }
 
 export function registerExtension(context: vscode.ExtensionContext) {
 	RootConfig = vscode.workspace.getConfiguration('msr');
-	context.subscriptions.push(vscode.languages.registerDefinitionProvider('*', new DefinitionFinder));
-	context.subscriptions.push(vscode.languages.registerReferenceProvider('*', new ReferenceFinder));
+	const selector = {
+		languageId: '*',
+		scheme: 'file',
+	};
 
-	context.subscriptions.push(vscode.commands.registerTextEditorCommand('msr.regexFindDefinition',
-		(textEditor: vscode.TextEditor, edit: vscode.TextEditorEdit, ...args: any[]) =>
-			runFindingCommand(MyCommandType.RegexFindDefinitionInCodeFiles, textEditor, edit, args)));
+	context.subscriptions.push(vscode.languages.registerDefinitionProvider(selector, new DefinitionFinder));
+	context.subscriptions.push(vscode.languages.registerReferenceProvider(selector, new ReferenceFinder));
 
-	context.subscriptions.push(vscode.commands.registerTextEditorCommand('msr.regexFindReference',
+	context.subscriptions.push(vscode.window.onDidCloseTerminal(terminal => {
+		if (terminal.name === RunCmdTerminalName) {
+			disposeTerminal();
+		}
+	}));
+
+	context.subscriptions.push(vscode.commands.registerTextEditorCommand('msr.regexFindDefinitionInCodeFiles',
 		(textEditor: vscode.TextEditor, edit: vscode.TextEditorEdit, ...args: any[]) =>
-			runFindingCommand(MyCommandType.RegexFindReferencesInCodeFiles, textEditor, edit, args)));
+			runFindingCommand(FindCommandType.RegexFindDefinitionInCodeFiles, textEditor, edit, args)));
+
+	context.subscriptions.push(vscode.commands.registerTextEditorCommand('msr.regexFindReferenceInCodeFiles',
+		(textEditor: vscode.TextEditor, edit: vscode.TextEditorEdit, ...args: any[]) =>
+			runFindingCommand(FindCommandType.RegexFindReferencesInCodeFiles, textEditor, edit, args)));
+
+	context.subscriptions.push(vscode.commands.registerTextEditorCommand('msr.regexFindPureReferenceInCodeFiles',
+		(textEditor: vscode.TextEditor, edit: vscode.TextEditorEdit, ...args: any[]) =>
+			runFindingCommand(FindCommandType.RegexFindPureReferencesInCodeFiles, textEditor, edit, args)));
 
 	context.subscriptions.push(vscode.commands.registerTextEditorCommand('msr.regexFindInDocFiles',
 		(textEditor: vscode.TextEditor, edit: vscode.TextEditorEdit, ...args: any[]) =>
-			runFindingCommand(MyCommandType.RegexFindReferencesInDocs, textEditor, edit, args)));
+			runFindingCommand(FindCommandType.RegexFindReferencesInDocs, textEditor, edit, args)));
 
 	context.subscriptions.push(vscode.commands.registerTextEditorCommand('msr.regexFindInConfigFiles',
 		(textEditor: vscode.TextEditor, edit: vscode.TextEditorEdit, ...args: any[]) =>
-			runFindingCommand(MyCommandType.RegexFindReferencesInConfigFiles, textEditor, edit, args)));
+			runFindingCommand(FindCommandType.RegexFindReferencesInConfigFiles, textEditor, edit, args)));
 
 	context.subscriptions.push(vscode.commands.registerTextEditorCommand('msr.regexFindInAllProjectFiles',
 		(textEditor: vscode.TextEditor, edit: vscode.TextEditorEdit, ...args: any[]) =>
-			runFindingCommand(MyCommandType.RegexFindReferencesInAllProjectFiles, textEditor, edit, args)));
+			runFindingCommand(FindCommandType.RegexFindReferencesInAllProjectFiles, textEditor, edit, args)));
 
 	context.subscriptions.push(vscode.commands.registerTextEditorCommand('msr.regexFindInCodeAndConfig',
 		(textEditor: vscode.TextEditor, edit: vscode.TextEditorEdit, ...args: any[]) =>
-			runFindingCommand(MyCommandType.RegexFindReferencesInCodeAndConfig, textEditor, edit, args)));
+			runFindingCommand(FindCommandType.RegexFindReferencesInCodeAndConfig, textEditor, edit, args)));
 
 	context.subscriptions.push(vscode.commands.registerTextEditorCommand('msr.regexFindInAllSmallFiles',
 		(textEditor: vscode.TextEditor, edit: vscode.TextEditorEdit, ...args: any[]) =>
-			runFindingCommand(MyCommandType.RegexFindReferencesInAllSmallFiles, textEditor, edit, args)));
+			runFindingCommand(FindCommandType.RegexFindReferencesInAllSmallFiles, textEditor, edit, args)));
+
+	context.subscriptions.push(vscode.commands.registerTextEditorCommand('msr.findSelectedPlainTextInCodeFiles',
+		(textEditor: vscode.TextEditor, edit: vscode.TextEditorEdit, ...args: any[]) =>
+			runFindingCommand(FindCommandType.FindPlainTextInCodeFiles, textEditor, edit, args)));
+
+	context.subscriptions.push(vscode.commands.registerTextEditorCommand('msr.findSelectedPlainTextInConfigFiles',
+		(textEditor: vscode.TextEditor, edit: vscode.TextEditorEdit, ...args: any[]) =>
+			runFindingCommand(FindCommandType.FindPlainTextInConfigFiles, textEditor, edit, args)));
+
+	context.subscriptions.push(vscode.commands.registerTextEditorCommand('msr.findSelectedPlainTextInDocFiles',
+		(textEditor: vscode.TextEditor, edit: vscode.TextEditorEdit, ...args: any[]) =>
+			runFindingCommand(FindCommandType.FindPlainTextInDocFiles, textEditor, edit, args)));
+
+	context.subscriptions.push(vscode.commands.registerTextEditorCommand('msr.findSelectedPlainTextInCodeAndConfigFiles',
+		(textEditor: vscode.TextEditor, edit: vscode.TextEditorEdit, ...args: any[]) =>
+			runFindingCommand(FindCommandType.FindPlainTextInConfigAndConfigFiles, textEditor, edit, args)));
 
 	context.subscriptions.push(vscode.commands.registerTextEditorCommand('msr.findSelectedPlainTextInAllProjectFiles',
 		(textEditor: vscode.TextEditor, edit: vscode.TextEditorEdit, ...args: any[]) =>
-			runFindingCommand(MyCommandType.FindPlainTextInAllProjectFiles, textEditor, edit, args)));
+			runFindingCommand(FindCommandType.FindPlainTextInAllProjectFiles, textEditor, edit, args)));
 
 	context.subscriptions.push(vscode.commands.registerTextEditorCommand('msr.findSelectedPlainTextInAllSmallFiles',
 		(textEditor: vscode.TextEditor, edit: vscode.TextEditorEdit, ...args: any[]) =>
-			runFindingCommand(MyCommandType.FindPlainTextInAllSmallFiles, textEditor, edit, args)));
+			runFindingCommand(FindCommandType.FindPlainTextInAllSmallFiles, textEditor, edit, args)));
 
 	context.subscriptions.push(vscode.commands.registerTextEditorCommand('msr.sortProjectFilesBySize',
 		(textEditor: vscode.TextEditor, edit: vscode.TextEditorEdit, ...args: any[]) =>
-			runFindingCommand(MyCommandType.SortProjectFilesBySize, textEditor, edit, args)));
+			runFindingCommand(FindCommandType.SortProjectFilesBySize, textEditor, edit, args)));
 
 	context.subscriptions.push(vscode.commands.registerTextEditorCommand('msr.sortProjectFilesByTime',
 		(textEditor: vscode.TextEditor, edit: vscode.TextEditorEdit, ...args: any[]) =>
-			runFindingCommand(MyCommandType.SortProjectFilesByTime, textEditor, edit, args)));
+			runFindingCommand(FindCommandType.SortProjectFilesByTime, textEditor, edit, args)));
 
 	context.subscriptions.push(vscode.commands.registerTextEditorCommand('msr.sortAllFilesBySize',
 		(textEditor: vscode.TextEditor, edit: vscode.TextEditorEdit, ...args: any[]) =>
-			runFindingCommand(MyCommandType.SortAllFilesBySize, textEditor, edit, args)));
+			runFindingCommand(FindCommandType.SortAllFilesBySize, textEditor, edit, args)));
 
 	context.subscriptions.push(vscode.commands.registerTextEditorCommand('msr.sortAllFilesByTime',
 		(textEditor: vscode.TextEditor, edit: vscode.TextEditorEdit, ...args: any[]) =>
-			runFindingCommand(MyCommandType.SortAllFilesByTime, textEditor, edit, args)));
+			runFindingCommand(FindCommandType.SortAllFilesByTime, textEditor, edit, args)));
 }
 
 // this method is called when your extension is deactivated
@@ -145,19 +178,31 @@ export class ReferenceFinder implements vscode.ReferenceProvider {
 	}
 }
 
+// Cannot avoid too frequent searching by mouse hover + click, because `Visual Studio Code` will not effect. So let VSCode solve this bug.
+function isTooFrequentSearch() {
+	const elapse = process.hrtime(lastSearchTime);
+	const ms = elapse[0] * 1000 + elapse[1] / 1000000;
+	lastSearchTime = process.hrtime();
+	return ms < 900;
+}
+
 function searchMatchedWords(findType: FindType, document: vscode.TextDocument, position: vscode.Position, token: vscode.CancellationToken, configKeyName: string, checkSkipTestPath: boolean) {
 	try {
+		const parsedFile = path.parse(document.fileName);
+		const extension = parsedFile.ext.replace(/^\./, '').toLowerCase() || 'default';
+		if (MyConfig.DisabledFileExtensionRegex.test(extension)) {
+			outputDebug('Disabled for `*.' + extension + '` file in configuration: `msr.disable.extensionPatterns`');
+			return Promise.reject();
+		}
+
 		const [currentWord, currentWordRange, currentText] = getCurrentWordAndText(document, position);
 		if (!checkSearchToolExists() || token.isCancellationRequested || currentWord.length < 2 || !currentWordRange) {
-			return Promise.resolve(null);
+			return Promise.reject();
 		}
 
 		clearOutputChannel();
-
-		const parsedFile = path.parse(document.fileName);
-		const extension = parsedFile.ext.replace(/^\./, '').toLowerCase() || 'default';
 		const mappedExt = FileExtensionToConfigExtMap.get(extension) || extension;
-		if (getConfig().IsDebug) {
+		if (MyConfig.IsDebug) {
 			outputDebug('mappedExt = ' + mappedExt + ' , languageId = ' + document.languageId + ' , file = ' + document.fileName);
 		}
 
@@ -173,7 +218,9 @@ function searchMatchedWords(findType: FindType, document: vscode.TextDocument, p
 			extraOptions = '--np test ' + extraOptions;
 		}
 
-		const searchPathOptions = getSearchPathOptions(mappedExt, FindType.Definition === findType, true);
+		const isFindDefinition = FindType.Definition === findType;
+		const useExtraSearchPaths = isFindDefinition || MyConfig.RootConfig.get('findReference.useExtraPaths') as boolean;
+		const searchPathOptions = getSearchPathOptions(mappedExt, isFindDefinition, useExtraSearchPaths);
 		let commandLine = 'msr ' + searchPathOptions + ' -f ' + filePattern + ' ' + searchOptions + ' ' + extraOptions;
 		if (MyConfig.DefaultMaxSearchDepth > 0 && !CheckMaxSearchDepthRegex.test(commandLine)) {
 			commandLine = commandLine.trim() + ' -k ' + MyConfig.DefaultMaxSearchDepth.toString();
@@ -217,37 +264,13 @@ function getMatchedLocationsAsync(findType: FindType, cmd: string, ranker: Searc
 			}
 
 			if (stdout) {
-				allResults = parseCommandOutput(stdout, ranker);
+				allResults = parseCommandOutput(stdout, findType, cmd, ranker);
 			}
 
 			if (stderr) {
-				const summary = GetSummaryRegex.exec(stderr);
-				if (!summary || /(\s+(Error|ERROR|WARN|error|BOM file)\s+|Jumped out)/.test(stderr)) {
-					outputDebug('\n' + stderr.replace(TrimCommandLineRegex, ''));
-				} else if (summary) {
-					outputInfo('\n' + stderr.replace(TrimCommandLineRegex, ''));
-				}
-
-				if (summary) {
-					const match = /^Matched (\d+) lines.*?read (\d+) lines.*?Used (\d+\.\d*) s/.exec(stderr);
-					if (match) {
-						const matchCount = parseInt(match[1]);
-						const lineCount = parseInt(match[2]);
-						const costSeconds = parseFloat(match[3]);
-						outputDebug('Got matched count = ' + matchCount + ' and time cost = ' + costSeconds + ' from summary.');
-						sumTimeCost(findType, costSeconds, lineCount);
-						if (matchCount < 1 && RootConfig.get('enable.useGeneralFindingWhenNoResults') as boolean) {
-							const findCmd = findType === FindType.Definition ? MyCommandType.RegexFindDefinitionInCodeFiles : MyCommandType.RegexFindDefinitionInCodeFiles;
-							runFindingCommandByCurrentWord(findCmd, ranker.currentWord, ranker.currentFile);
-							outputInfo('Will run general search, please check results of `MSR-RUN-CMD` channel in `TERMINAL` tab. Disable `msr.enable.useGeneralFindingWhenNoResults` if you do not want.');
-							outputInfo('Try extensive search if still no results: Click a word or select a text  -->  Press `Ctrl + Shift + P`  -->  Type `msr` + `find` to choose and search.');
-						}
-						else if (matchCount > 1 && costSeconds <= MyConfig.ReRunCmdInTerminalIfCostLessThan) {
-							outputInfo('Will re-run and show clickable + colorful results in `MSR-RUN-CMD` channel in `TERMINAL` tab. Decrease `msr.reRunSearchInTerminalIfCostLessThan` value if you do not want.');
-							runCommandInTerminal(cmd);
-						}
-					} else {
-						outputDebug('Failed to get time cost in summary.');
+				if (!findAndProcessSummary(false, stderr, findType, cmd, ranker)) {
+					if (/\bmsr\b.*?\s+not\s+/.test(stderr)) {
+						checkSearchToolExists(true, false);
 					}
 				}
 			}
@@ -258,6 +281,46 @@ function getMatchedLocationsAsync(findType: FindType, cmd: string, ranker: Searc
 		token.onCancellationRequested(() =>
 			killProcessTree(process.pid));
 	});
+}
+
+function findAndProcessSummary(skipIfNotMatch: boolean, cmdOutput: string, findType: FindType, cmd: string, ranker: SearchProperty): boolean {
+	const summary = GetSummaryRegex.exec(cmdOutput);
+	if (!summary && skipIfNotMatch) {
+		return false;
+	}
+
+	if (/(\s+|\d+m)WARN\b/.test(cmdOutput)) {
+		outputDebug('\n' + cmdOutput.replace(TrimCommandLineRegex, ''));
+	} else {
+		outputError('\n' + cmdOutput.replace(TrimCommandLineRegex, ''));
+	}
+
+	if (!summary) {
+		return false;
+	}
+
+	const match = /^Matched (\d+) lines.*?read (\d+) lines.*?Used (\d+\.\d*) s/.exec(cmdOutput);
+	if (match) {
+		const matchCount = parseInt(match[1]);
+		const lineCount = parseInt(match[2]);
+		const costSeconds = parseFloat(match[3]);
+		outputDebug('Got matched count = ' + matchCount + ' and time cost = ' + costSeconds + ' from summary.');
+		sumTimeCost(findType, costSeconds, lineCount);
+		if (matchCount < 1 && RootConfig.get('enable.useGeneralFindingWhenNoResults') as boolean) {
+			const findCmd = findType === FindType.Definition ? FindCommandType.RegexFindDefinitionInCodeFiles : FindCommandType.RegexFindDefinitionInCodeFiles;
+			runFindingCommandByCurrentWord(findCmd, ranker.currentWord, ranker.currentFile);
+			outputInfo('Will run general search, please check results of `MSR-RUN-CMD` channel in `TERMINAL` tab. Disable `msr.enable.useGeneralFindingWhenNoResults` if you do not want.');
+			outputInfo('Try extensive search if still no results: Click a word or select a text  -->  Press `Ctrl + Shift + P`  -->  Type `msr` + `find` to choose and search.');
+		}
+		else if (matchCount > 1 && costSeconds <= MyConfig.ReRunCmdInTerminalIfCostLessThan) {
+			outputInfo('Will re-run and show clickable + colorful results in `MSR-RUN-CMD` channel in `TERMINAL` tab. Decrease `msr.reRunSearchInTerminalIfCostLessThan` value if you do not want.');
+			runCommandInTerminal(cmd);
+		}
+	} else {
+		outputDebug('Failed to get time cost in summary.');
+	}
+
+	return true;
 }
 
 function sumTimeCost(findType: FindType, costSeconds: Number, lineCount: Number) {
@@ -278,18 +341,30 @@ function sumTimeCost(findType: FindType, costSeconds: Number, lineCount: Number)
 	}
 }
 
-function parseCommandOutput(stdout: string, ranker: SearchProperty): vscode.Location[] {
-	const matchedFileLines = stdout.split(/\r\n|\n\r|\n|\r/);
+function parseCommandOutput(stdout: string, findType: FindType, cmd: string, ranker: SearchProperty): vscode.Location[] {
+	let matchedFileLines = stdout.trimRight().split(/\r\n|\n\r|\n|\r/);
+	const summaryText = matchedFileLines.length > 0 && GetSummaryRegex.test(matchedFileLines[matchedFileLines.length - 1]) ? matchedFileLines[matchedFileLines.length - 1] : '';
+	if (summaryText.length > 0) {
+		matchedFileLines.pop();
+	}
 
 	let allResults: vscode.Location[] = [];
 	if (!MyConfig.NeedSortResults || matchedFileLines.length < 2) {
 		matchedFileLines.map(line => {
-			outputInfo(line);
-			let a = parseMatchedText(line, ranker)[1];
-			if (a) {
-				allResults.push(a);
+			const [score, location] = parseMatchedText(line, ranker);
+			if (location) {
+				allResults.push(location);
+				let sc = location.range.start;
+				let fileRowColumn = line.replace(':' + (sc.line + 1) + ':', ':' + (sc.line + 1) + ':' + sc.character);
+				outputInfo(fileRowColumn);
+			} else {
+				outputInfo(line);
 			}
 		});
+
+		if (summaryText.length > 0) {
+			findAndProcessSummary(true, summaryText, findType, cmd, ranker);
+		}
 
 		return allResults;
 	}
@@ -305,6 +380,7 @@ function parseCommandOutput(stdout: string, ranker: SearchProperty): vscode.Loca
 		if (!location) {
 			return;
 		}
+
 		if (maxScore < score) {
 			maxScore = score.valueOf();
 		}
@@ -368,6 +444,10 @@ function parseCommandOutput(stdout: string, ranker: SearchProperty): vscode.Loca
 
 	console.log('Count = ' + scoreList.length + ' , averageScore = ' + averageScore.toFixed(1) + ' , min = ' + scoreList[0].toFixed(1) + ' , max = ' + scoreList[scoreList.length - 1].toFixed(1)
 		+ ', removeThreshold = ' + removeThreshold.toFixed(1) + ' , removedCount = ' + removedCount + ' , scoreWordsText = ' + ranker.scoreWordsText);
+
+	if (summaryText.length > 0) {
+		findAndProcessSummary(true, summaryText, findType, cmd, ranker);
+	}
 
 	return allResults;
 }